requirements.txt
asgiref==3.8.1
asttokens==3.0.0
black==24.1.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
cryptography==44.0.2
decorator==5.2.1
Django==5.1.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.2.0
flake8==7.0.0
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
ipython==8.21.0
jedi==0.19.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.3.0
wcwidth==0.2.13
whitenoise==6.6.0
